+++ components/calendar/src/dangi.rs
@@ -266,7 +266,7 @@ impl Calendar for Dangi {
     }
 
     fn day_of_month(&self, date: &Self::DateInner) -> crate::types::DayOfMonth {
-        types::DayOfMonth(date.0 .0.day as u32)
+        types::DayOfMonth(date.0 .0.day)
     }
 
     fn day_of_year_info(&self, date: &Self::DateInner) -> crate::types::DayOfYearInfo {
@@ -492,7 +492,7 @@ mod test {
             expected_rel_iso: i32,
             expected_cyclic: u8,
             expected_month: u8,
-            expected_day: u32,
+            expected_day: u8,
         }
 
         let cases = [
