+++ ffi/capi/bindings/cpp/icu4x/IsoDate.hpp
@@ -37,11 +37,11 @@ namespace capi {
     
     uint16_t icu4x_IsoDate_day_of_year_mv1(const icu4x::capi::IsoDate* self);
     
-    uint32_t icu4x_IsoDate_day_of_month_mv1(const icu4x::capi::IsoDate* self);
+    uint8_t icu4x_IsoDate_day_of_month_mv1(const icu4x::capi::IsoDate* self);
     
     icu4x::capi::IsoWeekday icu4x_IsoDate_day_of_week_mv1(const icu4x::capi::IsoDate* self);
     
-    uint32_t icu4x_IsoDate_week_of_month_mv1(const icu4x::capi::IsoDate* self, icu4x::capi::IsoWeekday first_weekday);
+    uint8_t icu4x_IsoDate_week_of_month_mv1(const icu4x::capi::IsoDate* self, icu4x::capi::IsoWeekday first_weekday);
     
     icu4x::capi::WeekOf icu4x_IsoDate_week_of_year_mv1(const icu4x::capi::IsoDate* self, const icu4x::capi::WeekCalculator* calculator);
     
@@ -97,7 +97,7 @@ inline uint16_t icu4x::IsoDate::day_of_year() const {
   return result;
 }
 
-inline uint32_t icu4x::IsoDate::day_of_month() const {
+inline uint8_t icu4x::IsoDate::day_of_month() const {
   auto result = icu4x::capi::icu4x_IsoDate_day_of_month_mv1(this->AsFFI());
   return result;
 }
@@ -107,7 +107,7 @@ inline icu4x::IsoWeekday icu4x::IsoDate::day_of_week() const {
   return icu4x::IsoWeekday::FromFFI(result);
 }
 
-inline uint32_t icu4x::IsoDate::week_of_month(icu4x::IsoWeekday first_weekday) const {
+inline uint8_t icu4x::IsoDate::week_of_month(icu4x::IsoWeekday first_weekday) const {
   auto result = icu4x::capi::icu4x_IsoDate_week_of_month_mv1(this->AsFFI(),
     first_weekday.AsFFI());
   return result;
